Offload attendance calculation to parent class
