--- app/assets/javascripts/school_administrator_dashboard/dashboard_snapshots_page.jsx
@@ -1,22 +1,0 @@
-import SchoolAbsenceDashboard from './school_absence_dashboard.jsx';
-//const serializedData = this.props.serializedData;
-
-export default React.createClass({
-  displayName: 'DashboardSnapshotsPage',
-
-  PropTypes: {
-    attendanceData: React.PropTypes.object.isRequired
-  },
-
-  render: function() {
-    return (
-      <div className="DashboardSnapshotsPage">
-        <SchoolAbsenceDashboard
-          attendanceData={this.props.attendanceData}/>
-      </div>);
-  }
-});
-
-//Renders snapshots
-//Click on snapshot gets to specific view
-
