+++ app/assets/javascripts/school_administrator_dashboard/school_absence_dashboard.jsx
@@ -1,9 +1,7 @@
 import _ from 'lodash';
 import DashboardBarChart from './dashboard_bar_chart.jsx';
-//import GraphHelpers from '../helpers/graph_helpers.js';
 
 window.shared || (window.shared = {});
-const GraphHelpers = window.shared.GraphHelpers;
 
 const styles = {
   chartBox: {
@@ -34,7 +32,10 @@ export default React.createClass({
   displayName: 'SchoolAbsenceDashboard',
 
   propTypes: {
-    attendanceData: React.PropTypes.object.isRequired
+    schoolAttendance: React.PropTypes.object.isRequired,
+    schoolAttendanceMonths: React.PropTypes.array.isRequired,
+    homeRoomAttendance: React.PropTypes.object.isRequired,
+    homeRoomAttendanceMonths: React.PropTypes.array.isRequired
   },
 
   getInitialState: function() {
@@ -42,111 +43,6 @@ export default React.createClass({
       start_date: "2015-01-01",
       end_date: moment.utc().format("YYYY-MM-DD")
     };
-  },
-
-  //returns map of homerooms to absence events
-  absencesByHomeroom: function() {
-    //Group data by homeroom
-    const homeRoomGroups = _.groupBy(this.props.attendanceData.absences, 'homeroom');
-    let homeroomAbsences = {};
-    //Simplify homeroom data into map of homerooms to total homeroom absences along with homeroom size
-    Object.keys(homeRoomGroups).forEach( (key) => {
-      const size = homeRoomGroups[key].length;
-      var key2 = key;
-      if (key == "null") key2 = "No Homeroom";
-      homeroomAbsences[key2] = {
-        //Group events into day buckets
-        absences: this.eventsGroupedByDay(_.flattenDeep(_.map(homeRoomGroups[key], function(student) {
-          return student.absences;
-        }))),
-        size: size
-      };
-    });
-    return homeroomAbsences;
-  },
-
-  //Returns a list of homerooms along with dates and attendance percentage for each
-  attendanceByHomeroom: function () {
-    let homeroomDays = this.absencesByHomeroom();
-    Object.keys(homeroomDays).forEach((homeroom) => {
-      homeroomDays[homeroom].absences = this.averageDailyAttendance(homeroomDays[homeroom].absences, homeroomDays[homeroom].size);
-    });
-    return homeroomDays;
-  },
-
-  //Returns list of dates along with attendance percentage for all active students in school
-  attendanceBySchool: function() {
-    let schoolAbsences = {};
-    const homerooms = this.absencesByHomeroom();
-    Object.keys(homerooms).forEach((key) => {
-      this.mergeDateLists(schoolAbsences, homerooms[key].absences);
-    });
-    const schoolAttendance = this.averageDailyAttendance(schoolAbsences, this.totalStudents());
-    return schoolAttendance;
-  },
-
-  //Returns average monthly attendance givenn set of daily averages. To be used after applying date filters.
-  monthlyAttendanceBySchool: function () {
-    const schoolAttendance = this.attendanceBySchool();
-    let monthlyAttendance = {};
-    Object.keys(schoolAttendance).forEach((day) => {
-      if (monthlyAttendance[moment(day).date(1).format("YYYY-MM-DD")] === undefined) {
-        monthlyAttendance[moment(day).date(1).format("YYYY-MM-DD")] = [schoolAttendance[day]];
-      } else {
-        monthlyAttendance[moment(day).date(1).format("YYYY-MM-DD")].push(...schoolAttendance[day]);
-      }
-    });
-    return monthlyAttendance;
-  },
-
-  averageDailyAttendance: function(dailyAbsences, size) {
-    let averageDailyAttendance = {};
-    //mapping each homeroom to an array of day buckets containing all absences for each day w/in the homeroom
-    Object.keys(dailyAbsences).forEach((day) => {
-      averageDailyAttendance[day] = (size - dailyAbsences[day].length)/size*100;
-    });
-    return averageDailyAttendance;
-  },
-
-  mergeDateLists: function(list1, list2) {
-    Object.keys(list2).map((key) => {
-      if (list1[key] === undefined) list1[key] = list2[key];
-      else list1[key].push(...list2[key]);
-    });
-    return list1;
-  },
-
-  //takes array of events and groups by day on which they occurred
-  eventsGroupedByDay: function(events) {
-    return _.groupBy(events, (event) => {
-      return moment.utc(event.occurred_at).format("YYYY-MM-DD");
-    });
-  },
-
-  totalStudents: function() {
-    return this.props.attendanceData.absences.length;
-  },
-
-  createMonthBuckets: function (absences) {
-    const monthKeys = this.createMonthKeys();
-    return GraphHelpers.eventsToMonthBuckets(monthKeys, absences);
-  },
-
-  createMonthKeys: function() {
-    return GraphHelpers.monthKeys(moment.utc(), 12); //12 here just gives a year, should be change w/ slider state
-  },
-
-  createYearCategories: function() {
-    //not clear why to put this in chart yet
-    return GraphHelpers.yearCategories(this.createMonthKeys());
-  },
-
-  //Month buckets for monthly averages
-  //Day buckets for daily averages
-  getAttendancePercentage: function(buckets, students){
-    return _.map(buckets, function(bucket) {
-      return (students - bucket.length)/students*100;
-    });
   },
 
   getFirstDateIndex: function(dates, start_date) {
@@ -176,32 +72,33 @@ export default React.createClass({
   },
 
   renderMonthlyAbsenceChart: function() {
-    const schoolEvents = this.monthlyAttendanceBySchool();
-    const eventMonths = Object.keys(schoolEvents).sort();
+    const schoolAttendance = this.props.schoolAttendance;
+    const schoolAttendanceMonths = this.props.schoolAttendanceMonths;
     //Apply filter to object keys
-    let filteredDates = _.slice(eventMonths,
-                                this.getFirstDateIndex(eventMonths,this.state.start_date),
-                                this.getLastDateIndex(eventMonths,this.state.end_date)
+    let filteredDates = _.slice(schoolAttendanceMonths,
+                                this.getFirstDateIndex(schoolAttendanceMonths,this.state.start_date),
+                                this.getLastDateIndex(schoolAttendanceMonths,this.state.end_date)
                                 );
     let filteredAttendanceSeries = filteredDates.map( (month) => {
-      return _.sum(schoolEvents[month])/schoolEvents[month].length;
+      return _.sum(schoolAttendance[month])/schoolAttendance[month].length;
     });
     const categories = filteredDates.map((month) => moment.utc(month).format("YYYY-MM"));
     // const yearCategories = GraphHelpers.yearCategories(categories);
+
     return (
       <DashboardBarChart
         id = {'string'}
         categories = {categories}
         seriesData = {filteredAttendanceSeries}
-        monthsBack = {12}
+        monthsBack = {categories.length}
         titleText = {'Attendance (Percent)'}/>
     );
   },
 
   renderHomeroomAbsenceChart: function() {
-    const homerooms = this.attendanceByHomeroom();
+    const homeRoomAttendance = this.props.homeRoomAttendance;
     //Insert filtering here
-    var homeroomSeries = _.map(homerooms, (homeroom) => {
+    var homeroomSeries = _.map(homeRoomAttendance, (homeroom) => {
       var percentages = _.map(homeroom.absences);
       return _.sum(percentages)/percentages.length;
     });
@@ -210,7 +107,7 @@ export default React.createClass({
       <DashboardBarChart
         style = {styles.chartBox}
         id = {'string'}
-        categories = {Object.keys(homerooms)}
+        categories = {Object.keys(homeRoomAttendance)}
         seriesData = {homeroomSeries}
         monthsBack = {12}
         titleText = {'Attendance (Percent)'}/>
